Add support for annotated tags using the releaseCommitMessage
In the release process I am trying to support we create annotated tags
with a commit message that references the release issue number.

We don't create a new commit for this but when we build we include the
version calcualted at runtime by the plugin into the projects build
meta-data.

This patch modifies how a tag is set so that if the createAnnotatedTag
flag is set the releaseCommitMessage will be used for the tag's message.